Battle System Base
Prototype of battle system using skills and character stats. I trun off
dialog part of code

+++ TheDude/Assets/Scripts/ScripObj/Ability.cs
@@ -15,7 +15,7 @@ public class Ability : ScriptableObject
 
     public float heal;
 
-    public int costSp;
+    public float costSp;
 
     public List<Effects> efects;
